Revert to v2.2.0
This commit reverts all changes made since v2.2.0. This is done to update
Mellanox PMDs to their latest version by back-porting the related upstream
commits on top of DPDK v2.2.0.

+++ drivers/net/bnx2x/bnx2x_vfpf.h
@@ -115,7 +115,7 @@ struct vf_resc {
 	uint8_t num_mc_filters;
 	uint8_t permanent_mac_addr[ETH_ALEN];
 	uint8_t current_mac_addr[ETH_ALEN];
-	uint32_t pf_link_speed;
+	uint16_t pf_link_speed;
 	uint32_t pf_link_supported;
 };
